Paradigm checkpoint
It's almost finished, really. But it's not going so hot because it
turns out SUID doesn't work on scripts. I needed this because
`--bib --display` needs privileges. I will have to throw a few
curveballs to prevent too much configuration on the user's part.

I'm going to re-engineer this a tad.